refactor: move proto transformer to handler
