+++ core/user/user.go
@@ -4,9 +4,6 @@ package user
 import (
 	"context"
 	"time"
-
-	compassv1beta1 "github.com/odpf/compass/api/proto/odpf/compass/v1beta1"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 // User is a basic entity of a user
@@ -17,66 +14,6 @@ type User struct {
 	Provider  string    `json:"provider" diff:"-" db:"provider"`
 	CreatedAt time.Time `json:"-" diff:"-" db:"created_at"`
 	UpdatedAt time.Time `json:"-" diff:"-" db:"updated_at"`
-}
-
-// ToProto transforms struct with some fields only to proto
-func (d User) ToProto() *compassv1beta1.User {
-	if d.UUID == "" {
-		return nil
-	}
-
-	return &compassv1beta1.User{
-		Uuid:  d.UUID,
-		Email: d.Email,
-	}
-}
-
-// ToFullProto transforms struct with all fields to proto
-func (d User) ToFullProto() *compassv1beta1.User {
-	if d.UUID == "" {
-		return nil
-	}
-
-	var createdAtPB *timestamppb.Timestamp
-	if !d.CreatedAt.IsZero() {
-		createdAtPB = timestamppb.New(d.CreatedAt)
-	}
-
-	var updatedAtPB *timestamppb.Timestamp
-	if !d.UpdatedAt.IsZero() {
-		updatedAtPB = timestamppb.New(d.UpdatedAt)
-	}
-
-	return &compassv1beta1.User{
-		Id:        d.ID,
-		Uuid:      d.UUID,
-		Email:     d.Email,
-		Provider:  d.Provider,
-		CreatedAt: createdAtPB,
-		UpdatedAt: updatedAtPB,
-	}
-}
-
-// NewFromProto transforms proto to struct
-func NewFromProto(proto *compassv1beta1.User) User {
-	var createdAt time.Time
-	if proto.GetCreatedAt() != nil {
-		createdAt = proto.GetCreatedAt().AsTime()
-	}
-
-	var updatedAt time.Time
-	if proto.GetUpdatedAt() != nil {
-		updatedAt = proto.GetUpdatedAt().AsTime()
-	}
-
-	return User{
-		ID:        proto.GetId(),
-		UUID:      proto.GetUuid(),
-		Email:     proto.GetEmail(),
-		Provider:  proto.GetProvider(),
-		CreatedAt: createdAt,
-		UpdatedAt: updatedAt,
-	}
 }
 
 // Validate validates a user is valid or not
