+++ internal/server/v1beta1/tag.go
@@ -5,11 +5,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
 
 	compassv1beta1 "github.com/odpf/compass/api/proto/odpf/compass/v1beta1"
 	"github.com/odpf/compass/core/tag"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/structpb"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 var (
@@ -53,7 +56,7 @@ func (server *APIServer) GetTagByAssetAndTemplate(ctx context.Context, req *comp
 		return nil, internalServerError(server.logger, fmt.Sprintf("error finding a tag with asset and template: %s", err.Error()))
 	}
 
-	tagPB, err := tg.ToProto()
+	tagPB, err := tagToProto(tg)
 	if err != nil {
 		return nil, internalServerError(server.logger, err.Error())
 	}
@@ -86,7 +89,7 @@ func (server *APIServer) CreateTagAsset(ctx context.Context, req *compassv1beta1
 
 	var tagValues []tag.TagValue
 	for _, tvPB := range req.GetTagValues() {
-		tagValues = append(tagValues, tag.NewTagValueFromProto(tvPB))
+		tagValues = append(tagValues, tagValueFromProto(tvPB))
 	}
 
 	tagDomain := tag.Tag{
@@ -113,7 +116,7 @@ func (server *APIServer) CreateTagAsset(ctx context.Context, req *compassv1beta1
 
 	}
 
-	tagPB, err := tagDomain.ToProto()
+	tagPB, err := tagToProto(tagDomain)
 	if err != nil {
 		return nil, internalServerError(server.logger, err.Error())
 	}
@@ -147,7 +150,7 @@ func (server *APIServer) UpdateTagAsset(ctx context.Context, req *compassv1beta1
 
 	var tagValues []tag.TagValue
 	for _, tvPB := range req.GetTagValues() {
-		tagValues = append(tagValues, tag.NewTagValueFromProto(tvPB))
+		tagValues = append(tagValues, tagValueFromProto(tvPB))
 	}
 
 	tagDomain := tag.Tag{
@@ -169,7 +172,7 @@ func (server *APIServer) UpdateTagAsset(ctx context.Context, req *compassv1beta1
 		return nil, internalServerError(server.logger, fmt.Sprintf("error updating an asset's tag: %s", err.Error()))
 	}
 
-	tagPB, err := tagDomain.ToProto()
+	tagPB, err := tagToProto(tagDomain)
 	if err != nil {
 		return nil, internalServerError(server.logger, err.Error())
 	}
@@ -230,7 +233,7 @@ func (server *APIServer) GetAllTagsByAsset(ctx context.Context, req *compassv1be
 
 	var tagsPB []*compassv1beta1.Tag
 	for _, tg := range tags {
-		tgPB, err := tg.ToProto()
+		tgPB, err := tagToProto(tg)
 		if err != nil {
 			return nil, internalServerError(server.logger, err.Error())
 		}
@@ -240,4 +243,109 @@ func (server *APIServer) GetAllTagsByAsset(ctx context.Context, req *compassv1be
 	return &compassv1beta1.GetAllTagsByAssetResponse{
 		Data: tagsPB,
 	}, nil
+}
+
+// tagToProto convert domain to protobuf
+func tagToProto(t tag.Tag) (*compassv1beta1.Tag, error) {
+	var tagValuesPB []*compassv1beta1.TagValue
+	for _, tv := range t.TagValues {
+		tvPB, err := tagValueToProto(tv)
+		if err != nil {
+			return nil, err
+		}
+		tagValuesPB = append(tagValuesPB, tvPB)
+	}
+
+	return &compassv1beta1.Tag{
+		AssetId:             t.AssetID,
+		TemplateUrn:         t.TemplateURN,
+		TagValues:           tagValuesPB,
+		TemplateDisplayName: t.TemplateDisplayName,
+		TemplateDescription: t.TemplateDescription,
+	}, nil
+}
+
+// tagFromProto converts proto to tag.Tag
+func tagFromProto(pb *compassv1beta1.Tag) tag.Tag {
+	var tagValues []tag.TagValue
+	if pb.GetTagValues() != nil {
+		for _, tvPB := range pb.GetTagValues() {
+			tv := tagValueFromProto(tvPB)
+			tagValues = append(tagValues, tv)
+		}
+	}
+
+	return tag.Tag{
+		AssetID:             pb.GetAssetId(),
+		TemplateURN:         pb.GetTemplateUrn(),
+		TagValues:           tagValues,
+		TemplateDisplayName: pb.GetTemplateDisplayName(),
+		TemplateDescription: pb.GetTemplateDescription(),
+	}
+}
+
+// tagValueToProto convert domain to protobuf
+func tagValueToProto(tv tag.TagValue) (*compassv1beta1.TagValue, error) {
+	var err error
+	var createdAtPB *timestamppb.Timestamp
+	if !tv.CreatedAt.IsZero() {
+		createdAtPB = timestamppb.New(tv.CreatedAt)
+	}
+
+	var updatedAtPB *timestamppb.Timestamp
+	if !tv.UpdatedAt.IsZero() {
+		updatedAtPB = timestamppb.New(tv.UpdatedAt)
+	}
+
+	var fieldValuePB *structpb.Value
+	if tv.FieldValue != nil {
+		fieldValuePB, err = structpb.NewValue(tv.FieldValue)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return &compassv1beta1.TagValue{
+		FieldId:          uint32(tv.FieldID),
+		FieldValue:       fieldValuePB,
+		FieldUrn:         tv.FieldURN,
+		FieldDisplayName: tv.FieldDisplayName,
+		FieldDescription: tv.FieldDescription,
+		FieldDataType:    tv.FieldDataType,
+		FieldOptions:     tv.FieldOptions,
+		FieldRequired:    tv.FieldRequired,
+		CreatedAt:        createdAtPB,
+		UpdatedAt:        updatedAtPB,
+	}, nil
+}
+
+// tagValueFromProto converts proto to tag.TagValue
+func tagValueFromProto(pb *compassv1beta1.TagValue) tag.TagValue {
+	var createdAt time.Time
+	if pb.GetCreatedAt() != nil {
+		createdAt = pb.GetCreatedAt().AsTime()
+	}
+
+	var updatedAt time.Time
+	if pb.GetUpdatedAt() != nil {
+		updatedAt = pb.GetUpdatedAt().AsTime()
+	}
+
+	var fieldValue interface{}
+	if pb.GetFieldValue() != nil {
+		fieldValue = pb.GetFieldValue().AsInterface()
+	}
+
+	return tag.TagValue{
+		FieldID:          uint(pb.GetFieldId()),
+		FieldValue:       fieldValue,
+		FieldURN:         pb.GetFieldUrn(),
+		FieldDisplayName: pb.GetFieldDisplayName(),
+		FieldDescription: pb.GetFieldDescription(),
+		FieldDataType:    pb.GetFieldDataType(),
+		FieldOptions:     pb.GetFieldOptions(),
+		FieldRequired:    pb.GetFieldRequired(),
+		CreatedAt:        createdAt,
+		UpdatedAt:        updatedAt,
+	}
 }
